Grover Search

The Grover.jl module implements a Grover's search use-case. Assisted by Oracle.jl and Diffusion.jl, the appropriate chosen test state is marked by applying an $n$CZ gate, and appropriately amplified through the required number of iterations $r\approx \pi\sqrt{2^{n+1}}/4$.

API

QuantExQASM.Grover.run_grover! — Method

run_grover!(cct::Circuit.Circ, qubit_indices::Vector, state::Integer)

Generates a Grover search circuit sample, marking the state defined by state and performing iterations to amplify the desired result upon measurement.

QuantExQASM.Grover.mark_state! — Method

mark_state!(cct::Circuit.Circ, state::Integer, qubit_indices::Vector)

Applies the state marking procedure of the Grover iteration.

QuantExQASM.Grover.apply_grover_iteration! — Method

apply_grover_iteration!(cct::Circuit.Circ, qubit_indices::Vector)

Applies a single Grover iteration. To be used following mark_state!

QuantExQASM.Grover.state_init! — Method

state_init!(cct::Circuit.Circ, qubit_indices::Vector)

Initialises the state to the required target; defaults to $H^{\otimes n}\vert \psi \rangle$ . Override for custom initialisation.

QuantExQASM.Grover.calc_iterations — Method

calc_iterations(num_states::Integer)

Calculate the required number of iterations to maximise the state's amplitude.

QuantExQASM.Grover.apply_x! — Function

apply_x!(cct::Circuit.Circ, tgt, reg::Union{String, Nothing}=nothing)

Internally used PauliX implementation. Uses GateOps definition. Override for custom use.

QuantExQASM.Grover.apply_h! — Function

apply_h!(cct::Circuit.Circ, tgt, reg::Union{String, Nothing}=nothing)

Internally used Hadamard implementation. Uses GateOps definition. Override for custom use.

Example

To use the Grover module, we provide example code below to search for a state in a 5-qubit quantum register marked by bit-pattern 11 (0b01011).

using QuantExQASM

# Set 5-qubit limit on circuit
num_qubits = 5

# Set bit-pattern to 11 (0b01011)
bit_pattern = 11

# Do not use optimised routines
use_aux_qubits = false

# Create empty circuit with qiven qubit count
cct = QuantExQASM.Circuit.Circ(num_qubits)

# Initialise intermediate gates for use in NCU
QuantExQASM.NCU.init_intermed_gates(cct, num_qubits-1)

# Run Grover algorithm for given oracle bit-pattern
QuantExQASM.Grover.run_grover!(cct, collect(0:num_qubits-1), bit_pattern)

From here we can examine how many operations were generated as

println(cct.circ_ops.len)

815

We can also generate the OpenQASM equivalent circuit as

# Convert circuit to QASM populated buffer
cct_s = QuantExQASM.Circuit.to_qasm(cct, true)

# Print first 8 lines of QASM file
for i in split( String(cct_s), "\n")[1:8]
    println(i)
end

OPENQASM 2.0;
include "qelib1.inc";
qreg q[5];
creg c[5];

gate sssz tgt{
    u3(0.0,0.19634954084936204,0.19634954084936204) tgt;
}