NCU module

The NCU module enables the implementation of an $n$-qubit controlled unitary gate. In this instances, we have chosen to work primarily with $n$CX and $n$CZ gates, though others may also be used.

Implementation

The general principle follows the work of Barenco et al., Phys. Rev. A 52, 3457, (1995). All higher-order controlled quantum gates can be decomposed into 1 and 2 qubit gate calls, with varying degrees of optimisation. We implement the default quadratic gate decomposition strategy, alongside the 3CU optimised, and auxiliary register optimised variants.

API

QuantExQASM.NCU.apply_ncu! — Method

apply_ncu!(circuit::Circuit.Circ, q_ctrl::Vector, q_aux::Vector, q_tgt, U::GateOps.GateLabel)

Apply an n-qubit controlled gate operation on the given target qubit. Ensure the gate corresponding with symbol U is registered with g_map before use. Appends the GateCall operations to circuit

Arguments

circuit::Circuit.Circ

ctrls::Vector:

aux::Vector:

tgt::Int:

U::GateOps.GateLabel:

QuantExQASM.NCU.init_intermed_gates — Method

init_intermed_gates(circ::Circuit.Circ, num_ctrl::Union{Nothing, Int})

Generates the required intermediate gates for implementing the NCU using the given Circuit gate-set. Calling this routine is essential before running the NCU.apply_ncu! function.

QuantExQASM.NCU.register_gate — Method

register_gate(circ::Circuit.Circ, U::GateOps.GateLabel, gate::Matrix{<:Number})

Adds the user-defined gate to the cache with the pairing U=>gate

QuantExQASM.NCU.gen_intermed_gates — Method

gen_intermed_gates(ctrl_depth::Int, U::GateOps.GateLabel)

Generates all intermediate n-th root gates upto the given control depth.

QuantExQASM.NCU.get_intermed_gate — Method

get_intermed_gate(U::GateOps.GateLabel)

Generate sqrt U and adjoint sqrt U, given a previously cached gate U.

QuantExQASM.NCU.apply_cx! — Method

apply_cx!(c::Circuit.Circ, ctrl, tgt, reg)

Module-specific CX gate. Defaults to using the implementation from GateOps. Override for custom functionality.

QuantExQASM.NCU.apply_cu! — Method

apply_cu!(c::Circuit.Circ, ctrl, tgt, reg, gl::GateOps.GateLabel)

Module-specific CU gate. Defaults to using the implementation from GateOps. Override for custom functionality.

Example

To use the NCU module, we provide example code below to apply an n-controlled Pauli Z gate using either the optimised or unoptimised (quadratic) decomposition routines.

using QuantExQASM

# Set 5-qubit limit on circuit
num_qubits = 5

# Do not use optimised routines
use_aux_qubits = false

# Create Pauli-Z gate-label for application
gl = QuantExQASM.GateOps.GateLabel(:z)

# Create empty circuit with qiven qubit count
cct = QuantExQASM.Circuit.Circ(num_qubits)

# Initialise default intermediate gates (X,Y,Z,H) for use in NCU
QuantExQASM.NCU.init_intermed_gates(cct, num_qubits-1)

U = QuantExQASM.Circuit.gate_cache[gl]

if use_aux_qubits == true && (num_qubits/2+1 >= 4)
    ctrl = collect(0:Int( floor((num_qubits-1)//2) ))
    aux = collect(1 + Int( floor((num_qubits-1)//2)):num_qubits-2)
    tgt = num_qubits-1
else
    ctrl = collect(0:num_qubits-2)
    aux = Int[]
    tgt = num_qubits-1
end

for i in ctrl
    QuantExQASM.Circuit.add_gatecall!(cct, QuantExQASM.GateOps.pauli_x(i) )
end

QuantExQASM.NCU.apply_ncu!(cct, ctrl, aux, tgt, gl);

# Number of generation gate-call operations
println(cct.circ_ops.len)

73