Gate Operations

Gate type-system hierarchy

Abstraction of types for representing gates.

QuantExQASM.GateOps.AGate — Type

Abstract Gate type

QuantExQASM.GateOps.AGateLabel — Type

Abstract Gate Label, for unique gates

QuantExQASM.GateOps.AGateCall — Type

Abstract Gate Call, for tracking Gate labels applied to specific qubits

GateLabel and GateCall

Specific gate instance in a circuit, applied to a given (set of) qubit(s).

QuantExQASM.GateOps.GateLabelP — Type

Parametric Gate label. Tracks the gate symbol (:x,:y,:z, etc), and arbitrary parameters

QuantExQASM.GateOps.GateCall1P — Type

Parametric single qubit Gate call. Has a GateLabelP, target qubit index and register label

QuantExQASM.GateOps.GateCall2P — Type

Parametric two qubit Gate call. Has a GateLabelP, control and target qubit indices, register label, and base gate (assumes controlled U).

QuantExQASM.GateOps.GateCallNP — Type

Parametric n-qubit Gate call. Has a GateLabelP, control vector indices, target qubit index, register label, and base gate.

QuantExQASM.GateOps.Gate — Type

Wrap of Matrix to fit type system hierarchy

Single Qubit Gates: default

QuantExQASM.GateOps.pauli_x — Function

pauli_x(q_target::Int, register::Union{String, Nothing}=nothing)

Generate a single qubit Pauli-x GateCall (GateCall1P) applied to the target qubit (on given register, if provided)

Examples

julia> GateOps.pauli_x(0, "q")
QuantExQASM.GateOps.GateCall1P{Int64,Float64}(QuantExQASM.GateOps.GateLabelP{Int64,Float64}(:x, nothing), 0, "q")

QuantExQASM.GateOps.pauli_y — Function

pauli_y(q_target::Int, register::Union{String, Nothing}=nothing)

Generate a single qubit Pauli-y GateCall (GateCall1P) applied to the target qubit (on given register, if provided)

Examples

julia> GateOps.pauli_y(0, "q")
QuantExQASM.GateOps.GateCall1P{Int64,Float64}(QuantExQASM.GateOps.GateLabelP{Int64,Float64}(:y, nothing), 0, "q")

QuantExQASM.GateOps.pauli_z — Function

pauli_z(q_target::Int, register::Union{String, Nothing}=nothing)

Generate a single qubit Pauli-z GateCall (GateCall1P) applied to the target qubit (on given register, if provided)

Examples

julia> GateOps.pauli_z(0, "q")
QuantExQASM.GateOps.GateCall1P{Int64,Float64}(QuantExQASM.GateOps.GateLabelP{Int64,Float64}(:z, nothing), 0, "q")

QuantExQASM.GateOps.hadamard — Function

hadamard(q_target::Int, register::Union{String, Nothing}=nothing)

Generate a single qubit Hadamard GateCall (GateCall1P) applied to the target qubit (on given register, if provided)

Examples

julia> GateOps.hadamard(0, "q")
GateOps.GateCall1P{Int64,Float64}(GateOps.GateLabelP{Int64,Float64}(:h, nothing), 0, "q")

Single Qubit Gates: rotation and arbitrary

QuantExQASM.GateOps.u — Function

u(label::GateLabel, q_target::Int, register::Union{String, Nothing}=nothing)

Generate a single qubit arbitrary unitary GateCall (GateCall1P) applied to the target qubit (on given register, if provided)

Examples

julia> GateOps.u(GateOps.GateLabel(:mygate), 1, "qr")
GateOps.GateCall1P{Int64,Float64}(GateOps.GateLabelP{Int64,Float64}(:mygate, nothing), 1, "qr")

QuantExQASM.GateOps.r_x — Function

r_x(q_target::Int, theta::Real, register::Union{String, Nothing}=nothing)

Generate a single qubit R_x(θ) GateCall (GateCall1P) applied to the target qubit (on given register, if provided)

Examples

julia> GateOps.r_x(3,pi/2)
GateOps.GateCall1P{Int64,Float64}(GateOps.GateLabelP{Int64,Float64}(Symbol("r_x_angle=1.5707963267948966"), Dict{String,Union{Bool, Float64, Int64}}("angle"=>1.5708)), 3, nothing)

QuantExQASM.GateOps.r_y — Function

r_y(q_target::Int, theta::Real, register::Union{String, Nothing}=nothing)

Generate a single qubit R_y(θ) GateCall (GateCall1P) applied to the target qubit (on given register, if provided)

Examples

julia> GateOps.r_y(3,pi/2)
GateOps.GateCall1P{Int64,Float64}(GateOps.GateLabelP{Int64,Float64}(Symbol("r_y_angle=1.5707963267948966"), Dict{String,Union{Bool, Float64, Int64}}("angle"=>1.5708)), 3, nothing)

QuantExQASM.GateOps.r_z — Function

r_z(q_target::Int, theta::Real, register::Union{String, Nothing}=nothing)

Generate a single qubit R_z(θ) GateCall (GateCall1P) applied to the target qubit (on given register, if provided)

Examples

julia> GateOps.r_z(3,pi/2)
GateOps.GateCall1P{Int64,Float64}(GateOps.GateLabelP{Int64,Float64}(Symbol("r_z_angle=1.5707963267948966"), Dict{String,Union{Bool, Float64, Int64}}("angle"=>1.5708)), 3, nothing)

QuantExQASM.GateOps.r_phase — Function

r_phase(q_target::Int, theta::Real, register::Union{String, Nothing}=nothing)

Generate a single qubit phase shift GateCall ( diag(1, exp(1im*theta)) , GateCall1P) applied to the target qubit (on given register, if provided)

Examples

julia> GateOps.r_phase(3,pi/2)
GateOps.GateCall1P{Int64,Float64}(GateOps.GateLabelP{Int64,Float64}(Symbol("r_phase_angle=1.5707963267948966"), Dict{String,Union{Bool, Float64, Int64}}("angle"=>1.5708)), 3, nothing)

Two Qubit Gates: default

QuantExQASM.GateOps.c_pauli_x — Function

c_pauli_x(q_target::Int, q_ctrl::Int, register::Union{String, Nothing}=nothing)

Generate a controlled Pauli-x (two qubit) GateCall (GateCall2P), controlled on the index q_ctrl applied to the target q_target (on given register, if provided).

Examples

julia> GateOps.c_pauli_x(0, 1, "qreg")
GateOps.GateCall2P{Int64,Float64}(GateOps.GateLabelP{Int64,Float64}(:c_x, nothing), 0, 1, GateOps.GateCall1P{Int64,Float64}(GateOps.GateLabelP{Int64,Float64}(:x, nothing), 0, "qreg"), "qreg")

QuantExQASM.GateOps.c_pauli_y — Function

c_pauli_y(q_target::Int, q_ctrl::Int, register::Union{String, Nothing}=nothing)

Generate a controlled Pauli-y (two qubit) GateCall (GateCall2P), controlled on the index q_ctrl applied to the target q_target (on given register, if provided).

Examples

julia> GateOps.c_pauli_y(0, 1, "qreg")
GateOps.GateCall2P{Int64,Float64}(GateOps.GateLabelP{Int64,Float64}(:c_y, nothing), 0, 1, GateOps.GateCall1P{Int64,Float64}(GateOps.GateLabelP{Int64,Float64}(:y, nothing), 0, "qreg"), "qreg")

QuantExQASM.GateOps.c_pauli_z — Function

c_pauli_z(q_target::Int, q_ctrl::Int, register::Union{String, Nothing}=nothing)

Generate a controlled Pauli-x (two qubit) GateCall (GateCall2P), controlled on the index q_ctrl applied to the target q_target (on given register, if provided).

Examples

julia> GateOps.c_pauli_z(0, 1, "qreg")
GateOps.GateCall2P{Int64,Float64}(GateOps.GateLabelP{Int64,Float64}(:c_z, nothing), 0, 1, GateOps.GateCall1P{Int64,Float64}(GateOps.GateLabelP{Int64,Float64}(:z, nothing), 0, "qreg"), "qreg")

Two Qubit Gates: rotation and arbitrary

QuantExQASM.GateOps.c_u — Function

c_u(label::GateLabel, q_target::Int, q_ctrl::Int, gc::GateCall1, register::Union{String, Nothing}=nothing)

Generate a controlled unitary (two qubit) GateCall (GateCall2P), controlled on the index q_ctrl applied to the target q_target (on given register, if provided).

Examples

julia> GateOps.c_u(GateOps.GateLabel(:myCU), 0, 1, GateOps.pauli_x(0), "q")
GateOps.GateCall2P{Int64,Float64}(GateOps.GateLabelP{Int64,Float64}(:myCU, nothing), 0, 1, GateOps.GateCall1P{Int64,Float64}(GateOps.GateLabelP{Int64,Float64}(:x, nothing), 0, nothing), "q")

QuantExQASM.GateOps.c_r_x — Function

c_r_x(q_target::Int, q_ctrl::Int, theta::Real, register::Union{String, Nothing}=nothing)

Generate a controlled rotation about x (exp(iθσx/2)) GateCall (GateCall2P), controlled on the index `qctrlapplied to the targetq_target` (on given register, if provided).

Examples

julia> QuantExQASM.GateOps.c_r_x( 0, 1, pi/2, "q")
QuantExQASM.GateOps.GateCall2P{Int64,Float64}(QuantExQASM.GateOps.GateLabelP{Int64,Float64}(Symbol("c_r_x_angle=1.5707963267948966"), Dict{String,Union{Bool, Float64, Int64}}("angle" => 1.5707963267948966)), 0, 1, nothing, nothing)

QuantExQASM.GateOps.c_r_y — Function

c_r_y(q_target::Int, q_ctrl::Int, theta::Real, register::Union{String, Nothing}=nothing)

Generate a controlled rotation about y (exp(iθσy/2)) GateCall (GateCall2P), controlled on the index `qctrlapplied to the targetq_target` (on given register, if provided).

Examples

julia> QuantExQASM.GateOps.c_r_y( 0, 1, pi/3, "q")
QuantExQASM.GateOps.GateCall2P{Int64,Float64}(QuantExQASM.GateOps.GateLabelP{Int64,Float64}(Symbol("c_r_y_angle=1.0471975511965976"), Dict{String,Union{Bool, Float64, Int64}}("angle" => 1.0471975511965976)), 0, 1, nothing, nothing)

QuantExQASM.GateOps.c_r_z — Function

c_r_z(q_target::Int, q_ctrl::Int, theta::Real, register::Union{String, Nothing}=nothing)

Generate a controlled rotation about z (exp(iθσz/2)) GateCall (GateCall2P), controlled on the index `qctrlapplied to the targetq_target` (on given register, if provided).

Examples

julia> QuantExQASM.GateOps.c_r_z( 0, 1, pi/4, "q")
QuantExQASM.GateOps.GateCall2P{Int64,Float64}(QuantExQASM.GateOps.GateLabelP{Int64,Float64}(Symbol("c_r_z_angle=0.7853981633974483"), Dict{String,Union{Bool, Float64, Int64}}("angle" => 0.7853981633974483)), 0, 1, nothing, nothing)

QuantExQASM.GateOps.c_r_phase — Function

c_r_phase(q_target::Int, q_ctrl::Int, theta::Real, register::Union{String, Nothing}=nothing)

Generate a controlled phase rotation about (controlled [1 0; 0 exp(iθ)] GateCall (GateCall2P), controlled on the index q_ctrl applied to the target q_target (on given register, if provided).

Examples

julia> QuantExQASM.GateOps.c_r_phase( 0, 1, pi/7, "q")
QuantExQASM.GateOps.GateCall2P{Int64,Float64}(QuantExQASM.GateOps.GateLabelP{Int64,Float64}(Symbol("c_r_phase_angle=0.4487989505128276"), Dict{String,Union{Bool, Float64, Int64}}("angle" => 0.4487989505128276)), 0, 1, nothing, nothing)