Circuits

To compose a quantum algorithm, we need a structured manner for representing the gate calls, as defined in GateOps.md.

Circuit.jl

The Circuit.jl module allows us to structure the gate-calls into a quantum circuit representation. Circuits are appendable, pushable, and are represented under the hood as a doubly linked-list (CList from module CircuitList.jl), to allow ease of further optimisation and manipulation, if required.

The Circuit module also maintains a cached mapping from GateLabels to numerical matrices, wherein new gates created are stored for later use by other circuits. This memoization improves the performance of constructing circuits involving calculating sub-matrices (see the algorithms modules for examples).

QuantExQASM.Circuit.Circ — Type

Circ

Structure for ordered quantum circuit gate operations. Maintains MLL of gate-calls and a set of the gates used.

QuantExQASM.Circuit.Circ — Method

Circ()

Default constructor for empty circuit. Registers Pauli and Hadamard gates during initialisation.

QuantExQASM.Circuit.Circ — Method

Circ(num_qubits)

Circ constructor for given number of qubits circuit. Registers Pauli and Hadamard gates during initialisation.

QuantExQASM.Circuit.add_to_cache — Method

add_to_cache(label::GateOps.GateLabel, mat::Matrix{<:Number})

Adds a mapping between label=>mat for fast retrieval of gates in circuit generation.

Examples

julia> Circuit.add_to_cache(GateOps.GateLabel(:x), [0 1;1 0].+0im)

QuantExQASM.Circuit.add_gatecall! — Method

add_gatecall!(circ::Circ, gc::GateOps.AGateCall)

Adds the given gate call gc to the circuit at the end position.

Examples

julia> Circuit.add_gatecall!(circ, GateOps.paul_x(4)) #Apply Paulix to qubit index 4

QuantExQASM.Circuit.add_gatecall! — Method

add_gatecall!(circ::Circ, gc::GateOps.AGateCall)

Adds the given gate call gc to the circuit at the end position.

Examples

julia> Circuit.add_gatecall!(circ, GateOps.c_paul_x(3,4)) #Apply C_Paulix to qubit index 3, controlled on 4

QuantExQASM.Circuit.to_string — Method

to_string(circ::Circ)

Convert the circuit to a string intermediate representation

Examples

julia> Circuit.add_gatecall!(circ, GateOps.paul_x(4)) #Apply Paulix to qubit index 4

QuantExQASM.Circuit.gatelabel_to_qasm — Method

gatelabel_to_qasm(gl::GateOps.GateLabel)

Defines an OpenQASM gate from the given GateLabel and matched matrix

QuantExQASM.Circuit.gatecall_to_qasm — Method

gatecall_to_qasm(gc::GateOps.GateCall1)

Convert a single-qubit gate call (GateCall1) to OpenQASM

QuantExQASM.Circuit.gatecall_to_qasm — Method

gatecall_to_qasm(gc::GateOps.GateCall2)

Convert a two-qubit gate call (GateCall2) to OpenQASM

QuantExQASM.Circuit.add_header — Function

add_header(num_qubits::Int, reg::String="q", creg::String="c")

Returns the given OpenQASM header for number of quantum and classical bits.

QuantExQASM.Circuit.to_qasm — Function

to_qasm(circ::Circ, header::Bool=true, filename::Union{String, Nothing}=nothing)

Convert a given circuit to OpenQASM.